Move EasyCodingStandardTester to EasyCodingStandard
