+++ packages/easy-coding-standard/src/Testing/Contract/ConfigAwareInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\EasyCodingStandard\Testing\Contract;
+
+interface ConfigAwareInterface
+{
+    public function provideConfig(): string;
+}
