+++ packages/easy-coding-standard/src/Testing/Exception/ShouldNotHappenException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\EasyCodingStandard\Testing\Exception;
+
+use Exception;
+
+final class ShouldNotHappenException extends Exception
+{
+}
